# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentStatusListView": {
        "additionalProperties": false,
        "properties": {
          "agents": {
            "items": {
              "$ref": "#/components/schemas/AgentStatusView"
            },
            "title": "Agents",
            "type": "array"
          },
          "in_flight": {
            "default": false,
            "title": "In Flight",
            "type": "boolean"
          },
          "scene_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Scene Id",
            "type": "string"
          }
        },
        "required": [
          "scene_id",
          "agents"
        ],
        "title": "AgentStatusListView",
        "type": "object"
      },
...
      "CommandAck": {
        "additionalProperties": false,
        "description": "命令受理结果。\n\n``deduplicated=True`` 表示重放同一个幂等命令，**未**新增模型请求或记录\n（PRD 5.3、5.4）。``event_status`` 区分“已接受”与“已生效”。",
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "boolean"
          },
          "command": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ControlCommandType"
              },
              {
                "type": "null"
              }
            ]
          },
          "deduplicated": {
            "default": false,
            "title": "Deduplicated",
            "type": "boolean"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "event_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "event_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Status"
          },
          "pause_reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PauseReason"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "run_state": {
            "$ref": "#/components/schemas/RunState"
          }
        },
        "required": [
          "request_id",
          "accepted",
          "run_state"
        ],
        "title": "CommandAck",
        "type": "object"
      },
...
      "ControlCommandRequest": {
        "additionalProperties": false,
        "description": "控制命令（PRD 5.2、5.4：携带 ``request_id`` 以便幂等）。",
        "properties": {
          "command": {
            "$ref": "#/components/schemas/ControlCommandType"
          },
          "request_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "command"
        ],
        "title": "ControlCommandRequest",
        "type": "object"
      },
...
      "EventView": {
        "additionalProperties": false,
        "description": "单个事件（含状态，界面需区分已接受／已生效）。",
        "properties": {
          "event": {
            "$ref": "#/components/schemas/Event"
          },
          "status": {
            "$ref": "#/components/schemas/EventStatus"
          }
        },
        "required": [
          "event",
          "status"
        ],
        "title": "EventView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InjectEventRequest": {
        "additionalProperties": false,
        "description": "提交人工事件（PRD 4.3）。",
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "request_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "target_agent_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Agent Id"
          },
          "visibility": {
            "$ref": "#/components/schemas/EventVisibility"
          }
        },
        "required": [
          "request_id",
          "body",
          "visibility"
        ],
        "title": "InjectEventRequest",
        "type": "object"
      },
...
      "SceneCreateRequest": {
        "additionalProperties": false,
        "description": "创建场景（会话）。\n\n``agents`` 的 2～8 人约束在契约层即校验，因此越界请求得到 422 而不是\n进入业务层（PRD 1.2、9）。",
        "properties": {
          "agents": {
            "items": {
              "$ref": "#/components/schemas/AgentSpecRequest"
            },
            "maxItems": 8,
            "minItems": 2,
            "title": "Agents",
            "type": "array"
          },
          "background": {
            "default": "",
            "maxLength": 2000,
            "title": "Background",
            "type": "string"
          },
          "max_analysis_requests": {
            "anyOf": [
              {
                "maximum": 200.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Analysis Requests"
          },
          "max_role_requests": {
            "anyOf": [
              {
                "maximum": 200.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Role Requests"
          },
          "title": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "agents"
        ],
        "title": "SceneCreateRequest",
        "type": "object"
      },
      "SceneDetailView": {
        "additionalProperties": false,
        "description": "场景详情：场景 + 本场角色集合（含模板快照）。",
        "properties": {
          "agents": {
            "items": {
              "$ref": "#/components/schemas/SceneAgent"
            },
            "title": "Agents",
            "type": "array"
          },
          "locked": {
            "title": "Locked",
            "type": "boolean"
          },
          "scene": {
            "$ref": "#/components/schemas/Scene"
          }
        },
        "required": [
          "scene",
          "agents",
          "locked"
        ],
        "title": "SceneDetailView",
        "type": "object"
      },
      "SceneListView": {
        "additionalProperties": false,
        "properties": {
          "scenes": {
            "items": {
              "$ref": "#/components/schemas/SceneSummaryView"
            },
            "title": "Scenes",
            "type": "array"
          }
        },
        "required": [
          "scenes"
        ],
        "title": "SceneListView",
        "type": "object"
      },
...
      "ViewpointView": {
        "additionalProperties": false,
        "description": "角色视角（PRD 7.2）：与调用器**实际使用相同**的上下文选择结果。",
        "properties": {
          "agent_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Agent Id",
            "type": "string"
          },
          "agent_name": {
            "title": "Agent Name",
            "type": "string"
          },
          "cutoff_seq": {
            "minimum": 0.0,
            "title": "Cutoff Seq",
            "type": "integer"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "prompt_template_id": {
            "title": "Prompt Template Id",
            "type": "string"
          },
          "public_roster": {
            "items": {
              "type": "string"
            },
            "title": "Public Roster",
            "type": "array"
          },
          "scene_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Scene Id",
            "type": "string"
          },
          "visible_kinds": {
            "items": {
              "type": "string"
            },
            "title": "Visible Kinds",
            "type": "array"
          },
          "visible_seq": {
            "items": {
              "type": "integer"
            },
            "title": "Visible Seq",
            "type": "array"
          }
        },
        "required": [
          "scene_id",
          "agent_id",
          "agent_name",
          "prompt_template_id",
          "cutoff_seq",
          "public_roster",
          "visible_seq",
          "visible_kinds",
          "prompt"
        ],
        "title": "ViewpointView",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "SceneWeave（角色互动沙盒）后端。M00 提供工程基础与公共契约；M01 提供角色模板与场景配置；M02/M03 提供可见性、调度与模型适配；M04 提供运行控制、事件、时间线、角色视角与 SSE；M06 提供只读行为分析。",
    "title": "SceneWeave",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/scenes": {
      "get": {
        "operationId": "list_scenes_api_scenes_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneListView"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Scenes",
        "tags": [
          "m01-scenes"
        ]
      },
      "post": {
        "operationId": "create_scene_api_scenes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SceneCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneDetailView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Scene",
        "tags": [
          "m01-scenes"
        ]
      }
    },
...
    "/api/scenes/{scene_id}/agents/status": {
      "get": {
        "description": "每个本场角色的执行状态（PRD 7.1：角色卡只显示名称与执行状态）。\n\n只读观察：不调用模型，不产生任何剧情事实。",
        "operationId": "agent_status_api_scenes__scene_id__agents_status_get",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentStatusListView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Agent Status",
        "tags": [
          "m04-runtime"
        ]
      }
    },
...
    "/api/scenes/{scene_id}/agents/{agent_id}/viewpoint": {
      "get": {
        "description": "角色视角：与调用器**同一** ContextBuilder 的输出（PRD 7.2）。",
        "operationId": "viewpoint_api_scenes__scene_id__agents__agent_id__viewpoint_get",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewpointView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Viewpoint",
        "tags": [
          "m04-runtime"
        ]
      }
    },
...
    "/api/scenes/{scene_id}/commands": {
      "post": {
        "description": "控制命令：`START`／`STEP`／`PAUSE`／`RESUME`／`STOP`（幂等）。",
        "operationId": "run_command_api_scenes__scene_id__commands_post",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ControlCommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAck"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Command",
        "tags": [
          "m04-runtime"
        ]
      }
    },
    "/api/scenes/{scene_id}/events": {
      "get": {
        "description": "列出全部事件及其状态（界面区分已接受／已生效）。",
        "operationId": "list_events_api_scenes__scene_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EventView"
                  },
                  "title": "Response List Events Api Scenes  Scene Id  Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Events",
        "tags": [
          "m04-runtime"
        ]
      },
      "post": {
        "description": "提交人工事件；暂停时立即生效但不自动继续（PRD 4.3）。",
        "operationId": "inject_event_api_scenes__scene_id__events_post",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InjectEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAck"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Inject Event",
        "tags": [
          "m04-runtime"
        ]
      }
    },
    "/api/scenes/{scene_id}/state": {
      "get": {
        "description": "运行状态与预算（含是否有在途调用）。",
        "operationId": "run_state_api_scenes__scene_id__state_get",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStateView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run State",
        "tags": [
          "m04-runtime"
        ]
      }
    },
    "/api/scenes/{scene_id}/stream": {
      "get": {
        "description": "SSE：按 ``seq`` 补发已提交时间线，支持断线追赶与去重（PRD 第 8 节）。\n\n- 每条：``id: <seq>`` + ``event: timeline`` + JSON 数据；\n- 心跳为注释行，**不进入剧情**；\n- ``replay_limit`` 用于有界追赶（补发满 N 条后关闭，客户端可再连）。",
        "operationId": "stream_api_scenes__scene_id__stream_get",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "since_seq",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Since Seq",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "replay_limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 1000,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Replay Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream",
        "tags": [
          "m04-runtime"
        ]
      }
    },
    "/api/scenes/{scene_id}/summary": {
      "get": {
        "description": "实际调用计数与失败统计（PRD 1.2 的「实际调用计数」）。",
        "operationId": "scenario_summary_api_scenes__scene_id__summary_get",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioSummaryView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scenario Summary",
        "tags": [
          "m04-runtime"
        ]
      }
    },
    "/api/scenes/{scene_id}/timeline": {
      "get": {
        "description": "历史/时间线：**只读**已保存数据，不调用模型（PRD 5.4）。",
        "operationId": "timeline_api_scenes__scene_id__timeline_get",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "since_seq",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Since Seq",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Timeline",
        "tags": [
          "m04-runtime"
        ]
      }
    },
...
  }
}